requirements.txt
amqp==5.3.1
asgiref==3.11.1
billiard==4.2.4
celery==5.6.2
certifi==2026.2.25
channels==4.3.2
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cron-descriptor==1.4.5
defusedxml==0.7.1
Django==6.0.2
django-celery-beat==2.9.0
django-cors-headers==4.9.0
django-timezone-field==7.2.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
httpie==3.2.4
idna==3.11
kombu==5.6.2
markdown-it-py==4.0.0
mdurl==0.1.2
multidict==6.7.1
packaging==26.0
prompt_toolkit==3.0.52
Pygments==2.19.2
PyJWT==2.11.0
PySocks==1.7.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
redis==7.2.1
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.3.3
setuptools==82.0.0
six==1.17.0
sqlparse==0.5.5
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
